refactor: flip the variant suffix so the unmarked component file is the base ui source
the unmarked alias has meant base ui since #4844, but variant file naming still had the unmarked file carrying the radix flavor. the 13 component pairs (12 in components/assistant-ui plus the direction pair in components/ui/radix) and the 8 docs sample pairs now keep the base ui source in the unmarked <name>.tsx and the radix variant in <name>.radix.tsx, so casual edits and unmarked imports land on the flavor the product ships by default; docs live demos and templates resolve the base flavor at runtime as a consequence.

the registry build derives the radix tree from .radix.tsx overrides while the base tree reads unmarked files, and the four variant validators (content, trees differ, slot parity, export parity) are restructured to iterate the radix side's override list so they stay non vacuous now that the base pass performs no suffix lookup; the test helper no longer defaults the override list to every file, and new cases prove each validator still fires. the public output contract is unchanged: bare registry urls keep serving radix and /base/ keeps serving base, verified by a byte identical dist rebuild. sync-templates drops the suffix lookup since minimal syncs the now unmarked base files, and the docs preview machinery derives the radix twin for the flavor switcher instead of the base one.
