+++ apps/docs/components/docs/samples/assistant-modal.radix.tsx
@@ -4,7 +4,7 @@ import { forwardRef, useState } from "react";
 import { BotIcon, ChevronDownIcon } from "lucide-react";
 import { AssistantModalPrimitive } from "@assistant-ui/react";
 import { Thread } from "@/components/assistant-ui/thread";
-import { TooltipIconButton } from "@/components/assistant-ui/tooltip-icon-button.base";
+import { TooltipIconButton } from "@/components/assistant-ui/tooltip-icon-button.radix";
 import { SampleFrame } from "./sample-frame";
 
 export function AssistantModalSample() {
