+++ apps/docs/components/docs/samples/select.tsx
@@ -20,6 +20,36 @@ const fruits = [
   { value: "orange", label: "Orange" },
 ];
 
+const frameworks = [
+  { value: "react", label: "React" },
+  { value: "vue", label: "Vue" },
+  { value: "svelte", label: "Svelte" },
+];
+
+const backends = [
+  { value: "node", label: "Node.js" },
+  { value: "python", label: "Python" },
+];
+
+const northAmericaTimezones = [
+  { value: "est", label: "Eastern Standard Time (EST)" },
+  { value: "cst", label: "Central Standard Time (CST)" },
+  { value: "mst", label: "Mountain Standard Time (MST)" },
+  { value: "pst", label: "Pacific Standard Time (PST)" },
+];
+
+const europeTimezones = [
+  { value: "gmt", label: "Greenwich Mean Time (GMT)" },
+  { value: "cet", label: "Central European Time (CET)" },
+  { value: "eet", label: "Eastern European Time (EET)" },
+];
+
+const asiaTimezones = [
+  { value: "ist", label: "India Standard Time (IST)" },
+  { value: "cst_china", label: "China Standard Time (CST)" },
+  { value: "jst", label: "Japan Standard Time (JST)" },
+];
+
 export function SelectSample() {
   const [value, setValue] = useState("apple");
 
@@ -89,22 +119,31 @@ export function SelectGroupsSample() {
 
   return (
     <SampleFrame className="flex h-auto items-center justify-center p-6">
-      <SelectRoot value={value} onValueChange={setValue}>
+      <SelectRoot
+        value={value}
+        onValueChange={(value) => value !== null && setValue(value)}
+        items={[...frameworks, ...backends]}
+      >
         <SelectTrigger className="w-48">
           <SelectValue placeholder="Select a framework..." />
         </SelectTrigger>
         <SelectContent>
           <SelectGroup>
             <SelectLabel>Frontend</SelectLabel>
-            <SelectItem value="react">React</SelectItem>
-            <SelectItem value="vue">Vue</SelectItem>
-            <SelectItem value="svelte">Svelte</SelectItem>
+            {frameworks.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectGroup>
           <SelectSeparator />
           <SelectGroup>
             <SelectLabel>Backend</SelectLabel>
-            <SelectItem value="node">Node.js</SelectItem>
-            <SelectItem value="python">Python</SelectItem>
+            {backends.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectGroup>
         </SelectContent>
       </SelectRoot>
@@ -121,40 +160,58 @@ export function SelectVariantsSample() {
     <SampleFrame className="flex h-auto items-center justify-center gap-4 p-6">
       <div className="flex flex-col gap-2">
         <span className="text-muted-foreground text-xs">Outline (default)</span>
-        <SelectRoot value={outlineValue} onValueChange={setOutlineValue}>
+        <SelectRoot
+          value={outlineValue}
+          onValueChange={(value) => value !== null && setOutlineValue(value)}
+          items={frameworks}
+        >
           <SelectTrigger variant="outline" className="w-32">
             <SelectValue />
           </SelectTrigger>
           <SelectContent>
-            <SelectItem value="react">React</SelectItem>
-            <SelectItem value="vue">Vue</SelectItem>
-            <SelectItem value="svelte">Svelte</SelectItem>
+            {frameworks.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectContent>
         </SelectRoot>
       </div>
       <div className="flex flex-col gap-2">
         <span className="text-muted-foreground text-xs">Ghost</span>
-        <SelectRoot value={ghostValue} onValueChange={setGhostValue}>
+        <SelectRoot
+          value={ghostValue}
+          onValueChange={(value) => value !== null && setGhostValue(value)}
+          items={frameworks}
+        >
           <SelectTrigger variant="ghost" className="w-32">
             <SelectValue />
           </SelectTrigger>
           <SelectContent>
-            <SelectItem value="react">React</SelectItem>
-            <SelectItem value="vue">Vue</SelectItem>
-            <SelectItem value="svelte">Svelte</SelectItem>
+            {frameworks.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectContent>
         </SelectRoot>
       </div>
       <div className="flex flex-col gap-2">
         <span className="text-muted-foreground text-xs">Muted</span>
-        <SelectRoot value={mutedValue} onValueChange={setMutedValue}>
+        <SelectRoot
+          value={mutedValue}
+          onValueChange={(value) => value !== null && setMutedValue(value)}
+          items={frameworks}
+        >
           <SelectTrigger variant="muted" className="w-32">
             <SelectValue />
           </SelectTrigger>
           <SelectContent>
-            <SelectItem value="react">React</SelectItem>
-            <SelectItem value="vue">Vue</SelectItem>
-            <SelectItem value="svelte">Svelte</SelectItem>
+            {frameworks.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectContent>
         </SelectRoot>
       </div>
@@ -170,27 +227,39 @@ export function SelectSizesSample() {
     <SampleFrame className="flex h-auto items-center justify-center gap-4 p-6">
       <div className="flex flex-col gap-2">
         <span className="text-muted-foreground text-xs">Default</span>
-        <SelectRoot value={defaultValue} onValueChange={setDefaultValue}>
+        <SelectRoot
+          value={defaultValue}
+          onValueChange={(value) => value !== null && setDefaultValue(value)}
+          items={frameworks}
+        >
           <SelectTrigger size="default" className="w-36">
             <SelectValue />
           </SelectTrigger>
           <SelectContent>
-            <SelectItem value="react">React</SelectItem>
-            <SelectItem value="vue">Vue</SelectItem>
-            <SelectItem value="svelte">Svelte</SelectItem>
+            {frameworks.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectContent>
         </SelectRoot>
       </div>
       <div className="flex flex-col gap-2">
         <span className="text-muted-foreground text-xs">Small</span>
-        <SelectRoot value={smValue} onValueChange={setSmValue}>
+        <SelectRoot
+          value={smValue}
+          onValueChange={(value) => value !== null && setSmValue(value)}
+          items={frameworks}
+        >
           <SelectTrigger size="sm" className="w-36">
             <SelectValue />
           </SelectTrigger>
           <SelectContent>
-            <SelectItem value="react">React</SelectItem>
-            <SelectItem value="vue">Vue</SelectItem>
-            <SelectItem value="svelte">Svelte</SelectItem>
+            {frameworks.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectContent>
         </SelectRoot>
       </div>
@@ -203,29 +272,38 @@ export function SelectScrollableSample() {
 
   return (
     <SampleFrame className="flex h-auto items-center justify-center p-6">
-      <SelectRoot value={value} onValueChange={setValue}>
+      <SelectRoot
+        value={value}
+        onValueChange={(value) => value !== null && setValue(value)}
+        items={[...northAmericaTimezones, ...europeTimezones, ...asiaTimezones]}
+      >
         <SelectTrigger className="w-64">
           <SelectValue placeholder="Select a timezone..." />
         </SelectTrigger>
         <SelectContent>
           <SelectGroup>
             <SelectLabel>North America</SelectLabel>
-            <SelectItem value="est">Eastern Standard Time (EST)</SelectItem>
-            <SelectItem value="cst">Central Standard Time (CST)</SelectItem>
-            <SelectItem value="mst">Mountain Standard Time (MST)</SelectItem>
-            <SelectItem value="pst">Pacific Standard Time (PST)</SelectItem>
+            {northAmericaTimezones.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectGroup>
           <SelectGroup>
             <SelectLabel>Europe</SelectLabel>
-            <SelectItem value="gmt">Greenwich Mean Time (GMT)</SelectItem>
-            <SelectItem value="cet">Central European Time (CET)</SelectItem>
-            <SelectItem value="eet">Eastern European Time (EET)</SelectItem>
+            {europeTimezones.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectGroup>
           <SelectGroup>
             <SelectLabel>Asia</SelectLabel>
-            <SelectItem value="ist">India Standard Time (IST)</SelectItem>
-            <SelectItem value="cst_china">China Standard Time (CST)</SelectItem>
-            <SelectItem value="jst">Japan Standard Time (JST)</SelectItem>
+            {asiaTimezones.map((item) => (
+              <SelectItem key={item.value} value={item.value}>
+                {item.label}
+              </SelectItem>
+            ))}
           </SelectGroup>
         </SelectContent>
       </SelectRoot>
