+++ packages/ui/src/components/assistant-ui/accordion.tsx
@@ -1,7 +1,6 @@
 "use client";
 
-import type { ComponentProps } from "react";
-import { Accordion as AccordionPrimitive } from "radix-ui";
+import { Accordion as AccordionPrimitive } from "@base-ui/react/accordion";
 import { ChevronDownIcon } from "lucide-react";
 import { cva, type VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
@@ -24,8 +23,7 @@ function Accordion({
   className,
   variant,
   ...props
-}: ComponentProps<typeof AccordionPrimitive.Root> &
-  VariantProps<typeof accordionVariants>) {
+}: AccordionPrimitive.Root.Props & VariantProps<typeof accordionVariants>) {
   return (
     <AccordionPrimitive.Root
       data-slot="accordion"
@@ -36,10 +34,7 @@ function Accordion({
   );
 }
 
-function AccordionItem({
-  className,
-  ...props
-}: ComponentProps<typeof AccordionPrimitive.Item>) {
+function AccordionItem({ className, ...props }: AccordionPrimitive.Item.Props) {
   return (
     <AccordionPrimitive.Item
       data-slot="accordion-item"
@@ -47,7 +42,7 @@ function AccordionItem({
         "aui-accordion-item group/accordion-item",
         "group-data-[variant=default]/accordion:border-b group-data-[variant=default]/accordion:last:border-b-0",
         "group-data-[variant=outline]/accordion:border-b group-data-[variant=outline]/accordion:last:border-b-0",
-        "group-data-[variant=ghost]/accordion:data-[state=open]:bg-muted/50 group-data-[variant=ghost]/accordion:rounded-lg",
+        "group-data-[variant=ghost]/accordion:data-open:bg-muted/50 group-data-[variant=ghost]/accordion:rounded-lg",
         className,
       )}
       {...props}
@@ -59,7 +54,7 @@ function AccordionTrigger({
   className,
   children,
   ...props
-}: ComponentProps<typeof AccordionPrimitive.Trigger>) {
+}: AccordionPrimitive.Trigger.Props) {
   return (
     <AccordionPrimitive.Header className="flex">
       <AccordionPrimitive.Trigger
@@ -74,7 +69,7 @@ function AccordionTrigger({
         {...props}
       >
         {children}
-        <ChevronDownIcon className="text-muted-foreground pointer-events-none size-4 shrink-0 transition-transform duration-200 ease-out group-data-[state=open]/accordion-trigger:rotate-180" />
+        <ChevronDownIcon className="text-muted-foreground pointer-events-none size-4 shrink-0 transition-transform duration-200 ease-out group-data-panel-open/accordion-trigger:rotate-180" />
       </AccordionPrimitive.Trigger>
     </AccordionPrimitive.Header>
   );
@@ -84,11 +79,11 @@ function AccordionContent({
   className,
   children,
   ...props
-}: ComponentProps<typeof AccordionPrimitive.Content>) {
+}: AccordionPrimitive.Panel.Props) {
   return (
-    <AccordionPrimitive.Content
+    <AccordionPrimitive.Panel
       data-slot="accordion-content"
-      className="aui-accordion-content data-[state=closed]:animate-accordion-up data-[state=open]:animate-accordion-down overflow-hidden text-sm"
+      className="aui-accordion-content data-closed:animate-accordion-up data-open:animate-accordion-down overflow-hidden text-sm"
       {...props}
     >
       <div
@@ -101,7 +96,7 @@ function AccordionContent({
       >
         {children}
       </div>
-    </AccordionPrimitive.Content>
+    </AccordionPrimitive.Panel>
   );
 }
