+++ packages/ui/src/components/assistant-ui/badge.radix.tsx
@@ -1,8 +1,7 @@
 "use client";
 
 import type { ComponentProps } from "react";
-import { mergeProps } from "@base-ui/react/merge-props";
-import { useRender } from "@base-ui/react/use-render";
+import { Slot } from "radix-ui";
 import { cva, type VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
 
@@ -40,27 +39,29 @@ const badgeVariants = cva(
   },
 );
 
-export type BadgeProps = useRender.ComponentProps<"span"> &
-  VariantProps<typeof badgeVariants>;
+export type BadgeProps = ComponentProps<"span"> &
+  VariantProps<typeof badgeVariants> & {
+    asChild?: boolean;
+  };
 
-function Badge({ className, variant, size, render, ...props }: BadgeProps) {
-  const dataProps = {
-    "data-slot": "badge",
-    "data-variant": variant,
-    "data-size": size,
-  } as ComponentProps<"span">;
+function Badge({
+  className,
+  variant,
+  size,
+  asChild = false,
+  ...props
+}: BadgeProps) {
+  const Comp = asChild ? Slot.Root : "span";
 
-  return useRender({
-    defaultTagName: "span",
-    props: mergeProps<"span">(
-      dataProps,
-      {
-        className: cn(badgeVariants({ variant, size }), className),
-      },
-      props,
-    ),
-    render,
-  });
+  return (
+    <Comp
+      data-slot="badge"
+      data-variant={variant}
+      data-size={size}
+      className={cn(badgeVariants({ variant, size }), className)}
+      {...props}
+    />
+  );
 }
 
 export { Badge, badgeVariants };
