+++ packages/ui/src/components/assistant-ui/badge.tsx
@@ -1,7 +1,8 @@
 "use client";
 
 import type { ComponentProps } from "react";
-import { Slot } from "radix-ui";
+import { mergeProps } from "@base-ui/react/merge-props";
+import { useRender } from "@base-ui/react/use-render";
 import { cva, type VariantProps } from "class-variance-authority";
 import { cn } from "@/lib/utils";
 
@@ -39,29 +40,27 @@ const badgeVariants = cva(
   },
 );
 
-export type BadgeProps = ComponentProps<"span"> &
-  VariantProps<typeof badgeVariants> & {
-    asChild?: boolean;
-  };
+export type BadgeProps = useRender.ComponentProps<"span"> &
+  VariantProps<typeof badgeVariants>;
 
-function Badge({
-  className,
-  variant,
-  size,
-  asChild = false,
-  ...props
-}: BadgeProps) {
-  const Comp = asChild ? Slot.Root : "span";
+function Badge({ className, variant, size, render, ...props }: BadgeProps) {
+  const dataProps = {
+    "data-slot": "badge",
+    "data-variant": variant,
+    "data-size": size,
+  } as ComponentProps<"span">;
 
-  return (
-    <Comp
-      data-slot="badge"
-      data-variant={variant}
-      data-size={size}
-      className={cn(badgeVariants({ variant, size }), className)}
-      {...props}
-    />
-  );
+  return useRender({
+    defaultTagName: "span",
+    props: mergeProps<"span">(
+      dataProps,
+      {
+        className: cn(badgeVariants({ variant, size }), className),
+      },
+      props,
+    ),
+    render,
+  });
 }
 
 export { Badge, badgeVariants };
