+++ packages/ui/src/components/assistant-ui/message-timing.tsx
@@ -4,8 +4,9 @@ import { useMessageTiming } from "@assistant-ui/react";
 import {
   Tooltip,
   TooltipContent,
+  TooltipProvider,
   TooltipTrigger,
-} from "@/components/ui/radix/tooltip";
+} from "@/components/ui/tooltip";
 import { cn } from "@/lib/utils";
 import type { FC } from "react";
 
@@ -43,55 +44,61 @@ export const MessageTiming: FC<{
   if (timing?.totalStreamTime === undefined) return null;
 
   return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <button
-          type="button"
-          data-slot="message-timing-trigger"
-          aria-label="Message timing"
-          className={cn(
-            "text-muted-foreground hover:bg-accent hover:text-accent-foreground flex items-center rounded-md p-1 font-mono text-xs tabular-nums transition-colors",
-            className,
-          )}
+    <TooltipProvider>
+      <Tooltip>
+        <TooltipTrigger
+          render={
+            <button
+              type="button"
+              data-slot="message-timing-trigger"
+              aria-label="Message timing"
+              className={cn(
+                "text-muted-foreground hover:bg-accent hover:text-accent-foreground flex items-center rounded-md p-1 font-mono text-xs tabular-nums transition-colors",
+                className,
+              )}
+            />
+          }
         >
           {formatTimingMs(timing.totalStreamTime)}
-        </button>
-      </TooltipTrigger>
-      <TooltipContent
-        side={side}
-        sideOffset={8}
-        data-slot="message-timing-popover"
-        className="bg-popover text-popover-foreground rounded-lg border px-3 py-2 shadow-md [&_span>svg]:hidden!"
-      >
-        <div className="grid min-w-35 gap-1.5 text-xs">
-          {timing.firstTokenTime !== undefined && (
+        </TooltipTrigger>
+        <TooltipContent
+          side={side}
+          sideOffset={8}
+          data-slot="message-timing-popover"
+          className="bg-popover text-popover-foreground rounded-lg border px-3 py-2 shadow-md [&_span>svg]:hidden!"
+        >
+          <div className="grid min-w-35 gap-1.5 text-xs">
+            {timing.firstTokenTime !== undefined && (
+              <div className="flex items-center justify-between gap-4">
+                <span className="text-muted-foreground">First token</span>
+                <span className="font-mono tabular-nums">
+                  {formatTimingMs(timing.firstTokenTime)}
+                </span>
+              </div>
+            )}
             <div className="flex items-center justify-between gap-4">
-              <span className="text-muted-foreground">First token</span>
+              <span className="text-muted-foreground">Total</span>
               <span className="font-mono tabular-nums">
-                {formatTimingMs(timing.firstTokenTime)}
+                {formatTimingMs(timing.totalStreamTime)}
               </span>
             </div>
-          )}
-          <div className="flex items-center justify-between gap-4">
-            <span className="text-muted-foreground">Total</span>
-            <span className="font-mono tabular-nums">
-              {formatTimingMs(timing.totalStreamTime)}
-            </span>
-          </div>
-          {timing.tokensPerSecond !== undefined && (
+            {timing.tokensPerSecond !== undefined && (
+              <div className="flex items-center justify-between gap-4">
+                <span className="text-muted-foreground">Speed</span>
+                <span className="font-mono tabular-nums">
+                  {timing.tokensPerSecond.toFixed(1)} tok/s
+                </span>
+              </div>
+            )}
             <div className="flex items-center justify-between gap-4">
-              <span className="text-muted-foreground">Speed</span>
+              <span className="text-muted-foreground">Chunks</span>
               <span className="font-mono tabular-nums">
-                {timing.tokensPerSecond.toFixed(1)} tok/s
+                {timing.totalChunks}
               </span>
             </div>
-          )}
-          <div className="flex items-center justify-between gap-4">
-            <span className="text-muted-foreground">Chunks</span>
-            <span className="font-mono tabular-nums">{timing.totalChunks}</span>
           </div>
-        </div>
-      </TooltipContent>
-    </Tooltip>
+        </TooltipContent>
+      </Tooltip>
+    </TooltipProvider>
   );
 };
