+++ packages/ui/src/components/assistant-ui/model-selector.tsx
@@ -20,7 +20,7 @@ import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "@/components/ui/radix/popover";
+} from "@/components/ui/popover";
 import {
   Command,
   CommandEmpty,
@@ -29,8 +29,9 @@ import {
   CommandItem,
   CommandList,
   CommandSeparator,
-} from "@/components/ui/radix/command";
-import { RadioGroup as RadioGroupPrimitive } from "radix-ui";
+} from "@/components/ui/command";
+import { RadioGroup } from "@base-ui/react/radio-group";
+import { Radio } from "@base-ui/react/radio";
 
 export type ModelSelectorEffortOption = {
   id: string;
@@ -405,7 +406,7 @@ function ModelSelectorContent({
       align={align}
       sideOffset={sideOffset}
       className={cn(
-        "bg-popover/95 w-72 min-w-(--radix-popover-trigger-width) overflow-hidden rounded-xl p-0 shadow-lg backdrop-blur-sm",
+        "bg-popover/95 w-72 min-w-(--anchor-width) overflow-hidden rounded-xl p-0 shadow-lg backdrop-blur-sm",
         className,
       )}
       {...props}
@@ -606,26 +607,25 @@ function ModelSelectorEffort({
       {...props}
     >
       <span className="text-muted-foreground text-xs">{label}</span>
-      <RadioGroupPrimitive.Root
+      <RadioGroup
         value={effort ?? ""}
         onValueChange={setEffort}
-        orientation="horizontal"
         aria-label={typeof label === "string" ? label : "Reasoning effort"}
         className="flex items-center gap-0.5"
       >
         {efforts.map((option) => (
-          <RadioGroupPrimitive.Item
+          <Radio.Root
             key={option.id}
             value={option.id}
             className={cn(
               "focus-visible:ring-ring/50 text-muted-foreground hover:text-foreground rounded-md px-2 py-1 text-xs transition-colors outline-none focus-visible:ring-2",
-              "data-[state=checked]:bg-accent data-[state=checked]:text-accent-foreground data-[state=checked]:font-medium",
+              "data-checked:bg-accent data-checked:text-accent-foreground data-checked:font-medium",
             )}
           >
             {option.name}
-          </RadioGroupPrimitive.Item>
+          </Radio.Root>
         ))}
-      </RadioGroupPrimitive.Root>
+      </RadioGroup>
     </div>
   );
 }
