+++ packages/ui/src/components/assistant-ui/tooltip-icon-button.radix.tsx
@@ -1,14 +1,15 @@
 "use client";
 
 import { type ComponentPropsWithRef, forwardRef } from "react";
+import { Slot } from "radix-ui";
 
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { Button } from "@/components/ui/button";
+} from "@/components/ui/radix/tooltip";
+import { Button } from "@/components/ui/radix/button";
 import { cn } from "@/lib/utils";
 
 export type TooltipIconButtonProps = ComponentPropsWithRef<typeof Button> & {
@@ -21,24 +22,22 @@ export const TooltipIconButton = forwardRef<
   TooltipIconButtonProps
 >(({ children, tooltip, side = "bottom", className, ...rest }, ref) => {
   return (
-    <TooltipProvider>
+    <TooltipProvider delayDuration={0}>
       <Tooltip>
-        <TooltipTrigger
-          render={
-            <Button
-              variant="ghost"
-              size="icon"
-              {...rest}
-              className={cn(
-                "aui-button-icon size-6 p-1 active:scale-90",
-                className,
-              )}
-              ref={ref}
-            />
-          }
-        >
-          {children}
-          <span className="aui-sr-only sr-only">{tooltip}</span>
+        <TooltipTrigger asChild>
+          <Button
+            variant="ghost"
+            size="icon"
+            {...rest}
+            className={cn(
+              "aui-button-icon size-6 p-1 active:scale-90",
+              className,
+            )}
+            ref={ref}
+          >
+            <Slot.Slottable>{children}</Slot.Slottable>
+            <span className="aui-sr-only sr-only">{tooltip}</span>
+          </Button>
         </TooltipTrigger>
         <TooltipContent side={side}>{tooltip}</TooltipContent>
       </Tooltip>
